requirements.txt
slack-bolt==1.21.3
flask==3.1.0
apscheduler==3.10.4
pytz==2024.2
anthropic==0.49.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.165.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
